rename event-observer.ts to event-manager.ts
