+++ src/main/event-manager.ts
